backend-master/requirements.txt
Django~=4.1.5
djangorestframework~=3.14.0
Markdown~=3.4.1
django-filter~=22.1
Pillow~=9.4.0
drf-yasg~=1.21.4